+++ slack_time/methods/users.py
@@ -3,7 +3,7 @@ from typing import IO
 from typing import Union
 
 from requests import Response
-from slack_time.api import SlackAPI
+from slack_time import SlackAPI
 from slack_time.utils import cached_property
 from slack_time.utils import make_file