+++ CHANGELOG.md
@@ -10,3 +10,7 @@
     * slack_time/utils.py: fixing file upload bug
 
     * TODO.md: adding about making flake8 compliant
+
+    * slack_time/methods/__init__.py (SlackTime): moving to slack_time/__init__.py
+
+    * general re-arrangements of auxillary thing